fix(tests): repair test compile breakage from #79 + #109 squash overlap
The squash-merges of #107 (mempalace) and #115 (punctuation slider) each
landed against pre-#79/pre-#109 main respectively, but neither carried
forward fixes for test fixtures that touched both surfaces:

- SettingsRepositoryUiImpl gained palaceConfig + palaceApi constructor
  args from #79; SettingsRepositoryModesTest still called the old 3-arg
  ctor, and SettingsRepositoryPunctuationPauseTest (new in #109) did too.
- SettingsRepositoryUi gained Memory Palace methods (setPalaceHost,
  setPalaceApiKey, clearPalaceConfig, testPalaceConnection); the
  SettingsViewModelModesTest + SettingsViewModelPunctuationPauseTest +
  RealPlaybackControllerUiTest FakeSettingsRepo classes don't override
  them, so the test compile fails with "is not abstract".
- RealPlaybackControllerUiTest also still references the dropped
  PunctuationPause enum + setPunctuationPause method.

This PR:
- Extracts FakeAuth / FakeHydrator / makeFakePalaceConfig /
  makeFakePalaceApi / FakeSecrets into SettingsRepositoryTestSupport.kt
  so the three settings-repo tests share one source of truth as the
  repo's constructor surface evolves.
- Adds palace-stub overrides + 5-arg ctor calls to all three repo tests
  + the two viewmodel tests + RealPlaybackControllerUiTest.
- Migrates RealPlaybackControllerUiTest off the dropped PunctuationPause
  enum onto PUNCTUATION_PAUSE_DEFAULT_MULTIPLIER + the new
  setPunctuationPauseMultiplier method.

Verified: :feature:testDebugUnitTest :app:testDebugUnitTest
:core-playback:testDebugUnitTest all green.
