+++ app/src/test/kotlin/in/jphe/storyvox/data/SettingsRepositoryPunctuationPauseTest.kt
@@ -6,9 +6,6 @@ import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.floatPreferencesKey
 import androidx.datastore.preferences.core.stringPreferencesKey
-import `in`.jphe.storyvox.data.auth.SessionHydrator
-import `in`.jphe.storyvox.data.auth.SessionState
-import `in`.jphe.storyvox.data.repository.AuthRepository
 import `in`.jphe.storyvox.feature.api.PUNCTUATION_PAUSE_DEFAULT_MULTIPLIER
 import `in`.jphe.storyvox.feature.api.PUNCTUATION_PAUSE_LONG_MULTIPLIER
 import `in`.jphe.storyvox.feature.api.PUNCTUATION_PAUSE_MAX_MULTIPLIER
@@ -21,9 +18,6 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.cancel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.runTest
@@ -64,7 +58,13 @@ class SettingsRepositoryPunctuationPauseTest {
             scope = scope,
             produceFile = { file },
         )
-        repo = SettingsRepositoryUiImpl(store, FakeAuth(), FakeHydrator())
+        repo = SettingsRepositoryUiImpl(
+            store = store,
+            auth = FakeAuth(),
+            hydrator = FakeHydrator(),
+            palaceConfig = makeFakePalaceConfig(tempFolder.newFolder("palace_ds"), scope),
+            palaceApi = makeFakePalaceApi(),
+        )
     }
 
     @After
@@ -210,22 +210,5 @@ class SettingsRepositoryPunctuationPauseTest {
         }
     }
 
-    private class FakeAuth : AuthRepository {
-        private val state = MutableStateFlow<SessionState>(SessionState.Anonymous)
-        override val sessionState: StateFlow<SessionState> = state.asStateFlow()
-        override suspend fun captureSession(
-            cookieHeader: String,
-            userDisplayName: String?,
-            userId: String?,
-            expiresAt: Long?,
-        ) = Unit
-        override suspend fun clearSession() = Unit
-        override suspend fun cookieHeader(): String? = null
-        override suspend fun verifyOrExpire(): SessionState = SessionState.Anonymous
-    }
-
-    private class FakeHydrator : SessionHydrator {
-        override fun hydrate(cookies: Map<String, String>) = Unit
-        override fun clear() = Unit
-    }
+    // FakeAuth / FakeHydrator / palace fakes live in [SettingsRepositoryTestSupport].
 }
